LPS-52303 Move NIO mocking support to portal-test for better reuse purpose
